[refactor]: 가독성 향상을 위한 파일명 변경
